Step 3 of major refactor. - removed any remaining superfluous includes + const params and local variables all the things + some const ref for function arguments that are more complex types = switching algorithms to use ranges and / or range_views when possible when possible = reformatting pass on pretty much everything + everything that can be marked final should be marked final now = some general include sorting + simple getters are now almost universally on the header side, and marked const Note: not marked [[nodiscard]] yet, but that's the eventual goal + making some general stuff static
Note: some inconsistencies with {}, () and = initialization remain
Note: a lot of reformatting was done with Rider's automatic code cleanup, there may be some inconsistent usage of for loops with / without {} because of it. For loops with {} are preferred for everything that doesn't contain a oneliner. Example:
Good
for (auto i : container)
   print i

Bad
for (auto i : container1)
   for (auto j : container2)
   {
      print(j + container1.size() * i
   }

for (auto i : container1) should also have {}

Note: comment cleanup pass pending